requirements.txt
streamlit
langchain
langchain-community
openai
Python-dotenv
langchain-openai